Edit

See merge request s13-webmobile1-sub1/S13P11E106!32

+++ Back/src/main/java/com/e106/demolition_king/game/dto/ReportDto.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReportDto {
-    private Long reportSeq;
+    private Integer reportSeq;
     private String userUuid;
     private Integer singleTopBuilding;
     private Integer multiTopBuilding;
